+++ NEWS.md
@@ -1,5 +1,7 @@
 # ellmer (development version)
 
+* `batch_chat()` now supports `chat_google_gemini()` for batch processing via
+  the Gemini Developer API (@xmarquez, #914).
 * ellmer will now distinguish text content from thinking content while streaming, allowing downstream packages like shinychat to provide specific UI for thinking content (@simonpcouch, #909).
 * `chat_github()` now uses `chat_openai_compatible()` for improved compatibility, and `models_github()` now supports custom `base_url` configuration (@D-M4rk, #877).
 * `chat_ollama()` now contains a slot for `top_k` within the `params` argument (@frankiethull).